updated on 17-10-2017

+++ code/computational_geometry/2d_line_intersection/line.cpp
@@ -12,25 +12,25 @@
 using namespace std;
 
 struct vec2 {
-    float x;
-    float y;
-    vec2(float x, float y):x(x),y(y){}
+    double x;
+    double y;
+    vec2(double x, double y):x(x),y(y){}
 };
 
 /* return point of intersection, in parent's coordinate space of its parameters */
-vec2 intersection(float x1, float y1, float x2, float y2, float x3, float y3, float x4, float y4){
-    float x12 = x1 - x2;
-    float x34 = x3 - x4;
-    float y12 = y1 - y2;
-    float y34 = y3 - y4;
+vec2 intersection(double x1, double y1, double x2, double y2, double x3, double y3, double x4, double y4){
+    double x12 = x1 - x2;
+    double x34 = x3 - x4;
+    double y12 = y1 - y2;
+    double y34 = y3 - y4;
     
-    float c = x12 * y34 - y12 * x34;
+    double c = x12 * y34 - y12 * x34;
     
-    float a = x1 * y2 - y1 * x2;
-    float b = x3 * y4 - y3 * x4;
+    double a = x1 * y2 - y1 * x2;
+    double b = x3 * y4 - y3 * x4;
     
-    float x = (a * x34 - b * x12) / c;
-    float y = (a * y34 - b * y12) / c;
+    double x = (a * x34 - b * x12) / c;
+    double y = (a * y34 - b * y12) / c;
     return vec2(x,y);
 }
